-divided usecases to "features" -added aliases to layers and usecases -changes usecase call from usecase to repository call
