+++ src/application/repositories/versions-repository.ts
@@ -1,7 +1,7 @@
-import { Version } from "../../domain/version";
 import { Observable } from "rxjs";
 import { InjectionToken } from "@angular/core";
-import { VersionMessage } from "src/domain/version-message";
+import { Version } from "@domain/version";
+import { VersionMessage } from "@domain/version-message";
 
 export interface VersionsRepository {
   getVersionsList(): Observable<Version[]>;
