+++ src/application/usecases/change-active-version.ts
@@ -1,10 +1,10 @@
 import { Injectable } from "@angular/core";
-import { Usecase } from "../base/use-case";
-import { ReviewPageStateService } from "src/data/services/review-page-state.service";
+import { Usecase } from "@application/base/use-case";
+import { ReviewPageStateService } from "@data/services/review-page-state.service";
 
 @Injectable({ providedIn: "root" })
 export class ChangeActiveVersionUseCase implements Usecase {
-  constructor(private reviewPageState: ReviewPageStateService) {}
+  constructor(private reviewPageState: ReviewPageStateService) { }
   execute(versionId: number): void {
     this.reviewPageState.setState({ activeVersionId: versionId });
   }
