+++ src/application/usecases/get-state.ts
@@ -1,14 +1,14 @@
 import { Injectable } from "@angular/core";
-import { Usecase } from "../base/use-case";
-import { ReviewPageStateService } from "src/data/services/review-page-state.service";
+import { Usecase } from "@application/base/use-case";
+import { ReviewPageStateService } from "@data/services/review-page-state.service";
+import { ReviewPageState } from "@domain/review-page-state";
 import { Observable } from "rxjs";
-import { ReviewPageState } from "src/domain/review-page-state";
 
 @Injectable({ providedIn: "root" })
 export class GetStateUseCase
   implements Usecase<Observable<ReviewPageState>>
 {
-  constructor(private reviewPageState: ReviewPageStateService) {}
+  constructor(private reviewPageState: ReviewPageStateService) { }
   execute(): Observable<ReviewPageState> {
     return this.reviewPageState.state$;
   }
