+++ src/application/usecases/sync-session/join-session.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Usecase } from "../base/use-case";
+import { Usecase } from "../../base/use-case";
 import { SyncService } from "src/data/services/sync.service";
 import { ReviewPageStateService } from "src/data/services/review-page-state.service";
 
@@ -8,7 +8,7 @@ export class JoinSessionUseCase implements Usecase {
   constructor(
     private syncService: SyncService,
     private reviewPageStateService: ReviewPageStateService
-  ) {}
+  ) { }
   execute(): void {
     const cb = (event: MessageEvent) => {
       const { changes } = event.data;
