+++ src/application/usecases/sync-session/leave-session.ts
@@ -1,10 +1,10 @@
 import { Injectable } from "@angular/core";
-import { Usecase } from "../base/use-case";
+import { Usecase } from "../../base/use-case";
 import { SyncService } from "src/data/services/sync.service";
 
 @Injectable({ providedIn: "root" })
 export class LeaveSessionUseCase implements Usecase {
-  constructor(private syncService: SyncService) {}
+  constructor(private syncService: SyncService) { }
   execute(): void {
     this.syncService.stop();
   }
