+++ src/application/usecases/versions-list/remove-item-from-playlist.ts
@@ -1,13 +1,13 @@
 import { Injectable } from "@angular/core";
 import { take } from "rxjs";
 import { ReviewPageStateService } from "src/data/services/review-page-state.service";
-import { Usecase } from "../base/use-case";
+import { Usecase } from "../../base/use-case";
 
 @Injectable({ providedIn: "root" })
-export class RemoveItemFromPlaylist implements Usecase {
+export class RemoveItemFromPlaylistUseCase implements Usecase {
   constructor(
     private reviewPageState: ReviewPageStateService,
-  ) {}
+  ) { }
   execute(versionId: number): void {
     this.reviewPageState.state$.pipe(take(1)).subscribe((currentState) => {
       const newPlaylist = currentState.playlist.filter(
