+++ src/index.html
@@ -1,8 +1,11 @@
 <html>
-  <head>
-    <title>My app</title>
-  </head>
-  <body>
-    <my-app>Loading...</my-app>
-  </body>
+
+<head>
+  <title>Angular Clean Code</title>
+</head>
+
+<body>
+  <my-app>Loading...</my-app>
+</body>
+
 </html>
