+++ src/ui/pages/review-page/review-page.component.ts
@@ -1,15 +1,13 @@
 import { CommonModule } from "@angular/common";
 import { Component, inject } from "@angular/core";
+import { ReviewPageState } from "@domain/review-page-state";
+import { HeaderMenuComponent } from "@ui/components/header-menu/header-menu.component";
+import { RightPanelComponent } from "@ui/components/right-panel/right-panel.component";
+import { VersionsListComponent } from "@ui/components/versions-list/versions-list.component";
+import { VideoMenuComponent } from "@ui/components/video-menu/video-menu.component";
+import { WorkspaceComponent } from "@ui/components/workspace/workspace.component";
+import { GetStateUseCase } from "@usecases/get-state";
 import { Observable } from "rxjs";
-import { ChangeVersionUseCase } from "src/application/usecases/change-version";
-import { GetStateUseCase } from "src/application/usecases/get-state";
-import { ReviewPageState } from "src/domain/review-page-state";
-import { HeaderMenuComponent } from "../../components/header-menu/header-menu.component";
-import { RightPanelComponent } from "../../components/right-panel/right-panel.component";
-import { ChangeActiveVersionUseCase } from "src/application/usecases/change-active-version";
-import { VersionsListComponent } from "src/ui/components/versions-list/versions-list.component";
-import { VideoMenuComponent } from "src/ui/components/video-menu/video-menu.component";
-import { WorkspaceComponent } from "src/ui/components/workspace/workspace.component";
 
 @Component({
   selector: "review-page",
@@ -27,16 +25,4 @@ import { WorkspaceComponent } from "src/ui/components/workspace/workspace.compon
 })
 export class ReviewPage {
   state$: Observable<ReviewPageState> = inject(GetStateUseCase).execute();
-
-  constructor(
-    private changeVersionUseCase: ChangeVersionUseCase,
-    private changeActiveVersionUseCase: ChangeActiveVersionUseCase
-  ) {}
-
-  processVersionChange(id: number) {
-    this.changeVersionUseCase.execute(id);
-  }
-  public processActiveVersionChange(activeVersionId: number) {
-    this.changeActiveVersionUseCase.execute(activeVersionId);
-  }
 }
